refactor: isolate hook status tool
